Main
Wanted to solve the classic monte hall problem in as few of lines as possible. Not sure if it counts as a legitimate solution because the program doesn't actually run through the problem but more so just calculates the probability of winning if ones switches doors versus staying with their original choice.